+++ util.c
@@ -5,9 +5,9 @@
 
 //------------------------------------------------------------------------------
 
-void normalize(float *v)
+void normalize(double *v)
 {
-    float k = 1.f / sqrtf(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+    double k = 1.0 / sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 
     v[0] *= k;
     v[1] *= k;
@@ -17,7 +17,7 @@ void normalize(float *v)
 //------------------------------------------------------------------------------
 // 2-vector and 4-vector normalized midpoint vector calculators
 
-void mid2(float *m, const float *a, const float *b)
+void mid2(double *m, const double *a, const double *b)
 {
     m[0] = a[0] + b[0];
     m[1] = a[1] + b[1];
@@ -26,8 +26,8 @@ void mid2(float *m, const float *a, const float *b)
     normalize(m);
 }
 
-void mid4(float *m, const float *a, const float *b,
-                    const float *c, const float *d)
+void mid4(double *m, const double *a, const double *b,
+                    const double *c, const double *d)
 {
     m[0] = a[0] + b[0] + c[0] + d[0];
     m[1] = a[1] + b[1] + c[1] + d[1];
@@ -41,7 +41,7 @@ void mid4(float *m, const float *a, const float *b,
 
 float lerp1(float a, float b, float t)
 {
-    return b * t + a * (1 - t);
+    return b * t + a * (1.0f - t);
 }
 
 float lerp2(float a, float b, float c, float d, float s, float t)
@@ -51,30 +51,30 @@ float lerp2(float a, float b, float c, float d, float s, float t)
 
 //------------------------------------------------------------------------------
 // 1D and 2D spherical linear interpolation
-
-void slerp1(float *a, const float *b, const float *c, float t)
+/*
+void slerp1(double *a, const double *b, const double *c, double t)
 {
-    const float k = acosf(b[0] * c[0] + b[1] * c[1] + b[2] * c[2]);
-    const float s = 1.f - t;
-    const float u = sinf(s * k) / sinf(k);
-    const float v = sinf(t * k) / sinf(k);
+    const double k = acosf(b[0] * c[0] + b[1] * c[1] + b[2] * c[2]);
+    const double s = 1.f - t;
+    const double u = sinf(s * k) / sinf(k);
+    const double v = sinf(t * k) / sinf(k);
 
     a[0] = b[0] * u + c[0] * v;
     a[1] = b[1] * u + c[1] * v;
     a[2] = b[2] * u + c[2] * v;
 }
 
-void slerp2(float *v, const float *a, const float *b,
-                      const float *c, const float *d, float x, float y)
+void slerp2(double *v, const double *a, const double *b,
+                      const double *c, const double *d, double x, double y)
 {
-    float t[3];
-    float u[3];
+    double t[3];
+    double u[3];
 
     slerp1(t, a, b, x);
     slerp1(u, c, d, x);
     slerp1(v, t, u, y);
 }
-
+*/
 //------------------------------------------------------------------------------
 
 int extcmp(const char *name, const char *ext)
